requirements.txt
numpy==1.23.4
Pillow==9.3.0
Pillow==9.5.0
streamlit==1.13.0
tensorflow==2.11.0
